Add support for ANDROID_USER_HOME
This is meant to ANDROID_SDK_HOME and ANDROID_PREFS_ROOT, and
match ANDROID_HOME (which is the sdk location).

Unlike the previous variable, this is the path to the folder to
use, rather than the folder in which the .android folder is created.

Similarly to the handling of _SDK_HOME and _PREFS_ROOT, the new code
checks for all 3 values and complains if more than 1 are set with
different values.

Bug: N/A
Test: updated tests and added new ones.
Change-Id: I74606e5cfa01f0eda388b1ca7e93c2ec52cddfdc