Support dynamic features in the new lint integration
This change adds some support for dynamic features in the new lint
integration. If an app has dynamic features, the dynamic features'
source providers are merged with the app's source providers when
creating the app's lint model for the lint task.

With this change, dynamic features give lint a checkDependencies-like
input - the lint task has an input file that points it to other input
files that Gradle doesn't know about. This means we need to mark the
lint task as never up-to-date in that case, as otherwise it will be
incorrectly up to date when the dynamic feature code it analyses
changes.

Bug: 160392650
Test: DynamicAppLintTest
Change-Id: I6dba969f22dacf38eb394383d4ee7063c5d6c30e