Fixes to support multiple schema versions
So far we've only had one version of the new-style schema, so there were
some places where we didn't correctly support having multiple versions.
There are three main changes here:
- In some places we were creating the latest versions of elements rather
than a version corresponding to the parent. Fixed to create the right
version. This also meant changing the default Channel from an object (of
the latest version) to an id, which can be used to create an object of
the appropriate version.
- Allowing SdkHandler to fall back to previous versions, if the latest
isn't available yet. This is useful during the transition to a new
schema version, where the new version might not yet have been published
(or available in your local cache, or whatever).
- Worked around a bug in xml catalog lookup during schema code
generation (not during studio runtime) where lookup gets NPE if there
isn't a systemId specified, even though lookup works fine using the
publicId with no systemId.

Test: pending
Bug: 177929448
Change-Id: I5e894d31cb664b7d47ae79d4a16cb48e4241b74f