Add soft attribute for dependencies
A "soft" dependency is one where we only download a new version of the
dependency if a previous version already installed. E.g. if A depends on
B non-softly and you install A, B will also always be installed. If A
softly depends on B v2 and we have B v1 installed, B will be updated to
v2 when we install A, but if B isn't installed at all, it won't be
installed when we install A.

Test: pending
Bug: none
Change-Id: I57e5ee5a488c1f472fba90c3393073a0da205b39