Convert all fake Android classes with static to Java
I'm not sure why, but occasionally (and inconsistently?) the Kotlin
compiler generates statics in a different way than what the original
Java would have generated, for example a Companion object or an INSTANCE
field.

This is partly because if you have Java code calling Kotlin, it will
call the correct path, but if you have Kotlin code calling Kotlin, even
if it's annotated with annotations like @JvmStatic, it seems to
(sometimes?) prefer the Kotlin path.

To be 100% safe, we should just avoid trying to emulate Android Java
classes which define statics we depend on with Kotlin versions.

Bug: N/A
Test: Covered by existing
Change-Id: Ia489e0f9e24399e4e3c2874e2911e8d0ecc40f30