Set up ExtractProguardFiles task dependency implicitly
Currently, ExtractProguardFiles task is set explicitly to be a
dependency of variant preBuild task. This is causing issue when this
task runs with clean task in a build with native project.

This CL fixes the above problem by setting up task dependencies
implicitly via input/output property and also converts the task to
kotlin.

Bug: 179062268
Test: existing + CMakeBasicProjectTest
Change-Id: I4cd8beb2f605d7287c42277783a94423e1d9a3e2