New fixture to handle delta model snapshot
Many model tests validate slight tweaks to a default
project state, and we want to be able to look only at the
differences. Snapshots are routinely 1,000+ lines and so we don't
want to always have full snapshots. This has 2 benefits:
- we can more easily ensure that the snapshot contains
  valid data based on the project modifications
- changes to other areas of the model do not require an
  update to all the snapshots.

This change completely rewrites the classes handling
snapshotting a model into a string/file.

The DumpBuilder class is rewritten and split in 3 different classes
- SnapshotItemRegistrar just holds the items of the snapshot
- SnapshotItemWriter writes the items into a string
- ModelSnapshotter is the front end to the snapshot code.
  It handles comparison to a reference model as well as normalization
  of the paths (difference with the reference model must be done
  on the normalized paths)

This change also introduces ReferenceModelComparator. This is a
base class for tests that need to create a project for which
we want to query and test the model, as well as its reference
project (and model).

Bug: N/A
Test: Updated existing tests to use the new infra and added unit tests for the new infra
Change-Id: Id20d7502084aac269d9af7ac07ee413cdbbf7bc2