[Partial Lint Analysis] Improve dynamic feature lint modeling
With this change, the base app shows up as a module dependency instead
of an external library dependency in the dynamic feature lint model.
This will allow lint to include issues from dynamic features in the
base app module's lint reports (after some changes in lint).

With this change, dynamic feature lint reports will include issues from
the base module when checkDependencies is true in the dynamic feature.

Feature-on-feature dependencies are still not fully-supported by lint;
that will be addressed in a subsequent change.

Bug: 178810169
Bug: 180672373
Test: LintDynamicFeatureTest
Change-Id: Idcb112d816c195985118c8753dd3397ff8f08222