Simple timer class for C/C++ build side
This implements a utility class so that you can write:
  time("operation") {
   // Long running
  }
And have the timings written out to a file in the build/ folder. Example,
  test
    operation 55ms
  test completed in 56ms

It honors nesting of the time scope and tracks and reports "gap" time that is missed when inner scopes don't cover 100% of outer. It's optimized ease of use and the file format is mean for human readability and isn't intended for fine-grained timing or cases where the timings need to be machine readable. For example, timings that appear trivial (under 10ms) are dropped from the output to produce a cleaner big picture.
That said, the 'time' API itself is amenable to switching from text file to .json.gz in the format understood by Perfetto and similar tools.

This CL only adds a few timing scopes to production code. I plan to add more as part of future perf investigations.

Bug: NA
Test: Unittested
Change-Id: Ic74d6d4608b48fd8efcf8ca4c22b165a9d4f75db