Model missing inputs to Lint task
Lint automatically looks for custom lint rule jars by
checking two pieces of global state: an environment
variable (ANDROID_LINT_JARS) and the contents of the
directory .android/lint.

This change adds a new @Classpath property on the
Lint task which tracks the latter (.android/lint) for
up-to-date checking. The same will be done for the
mentioned environment variable via a future CL.

Bug: 181746060
Bug: 181590927
Test: LintGlobalRuleJarsTest.kt
Change-Id: I34596e79348f2d5cdd3583991afad1fa04fb621b